+++ src/result/resultWrap.js
