--- src/updateElement.js
@@ -1,6 +1,0 @@
-export function ResultElement(timeEle,time) {  
-    let $timeEle = timeEle;
-    //갱신 시간
-    $timeEle.innerHTML = `${time}`
-}
-
